Getting to updating examples.
